api/services/login/requirements.txt
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==40.0.1
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.95.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
pytest==7.2.2
python-jose==3.3.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.21.1
